File upload download

+++ controllers/admin.js
@@ -1,6 +1,8 @@
+const path = require('path');
 
 const { deleteById } = require('../models/product');
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 
 const { validationResult } = require('express-validator/check');
 
@@ -34,10 +36,32 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
   const errors = validationResult(req);
+  console.log(image);
+  //as image wasn't able to be read so re-rendering the page
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      errors: true,
+      formsCSS: true,
+      productCSS: true,
+      activeAddProduct: true,
+      errorMessage: 'image was not  of valid format',
+      olderInput: {
+        title: title,
+        price: price,
+        description: description,
+      },
+      validatorErrors: []
+    });
+  }
+
+  //as validation failed so re-renderring the page
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
@@ -50,20 +74,21 @@ exports.postAddProduct = (req, res, next) => {
       errorMessage: errors.array()[0].msg,
       olderInput: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description,
       },
       validatorErrors: errors.array()
     });
   }
-  const product = new Product({ title: title, price: price, description: description, imageUrl: imageUrl, userId: req.session.user._id });
+  const product = new Product({ title: title, price: price, description: description, imageUrl: image.filename, userId: req.session.user._id });
   product.save()   //save is the method by mongoose for that model
     .then((result) => {    //although save method doesn't returns a promise but gives a then & catch bloc
       console.log('product added');
       res.redirect('/admin/products');
     })
-    .catch((err) => { error500(err, next); });
+    .catch((err) => {
+      error500(err, next);
+    });
 };
 
 
@@ -100,7 +125,7 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.id;
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
   console.log(prodId);
@@ -120,7 +145,6 @@ exports.postEditProduct = (req, res, next) => {
           errorMessage: errors.array()[0].msg,
           olderInput: {
             title: title,
-            imageUrl: imageUrl,
             price: price,
             description: description,
           },
@@ -132,9 +156,12 @@ exports.postEditProduct = (req, res, next) => {
       else if (product.userId.toString() !== req.session.user._id.toString()) {
         return res.redirect('/');  //the product is not created by the current user, thus redirect
       }
+      if (image) {
+        fileHelper.deleteFile(path.join('images', product.imageUrl));
+        product.imageUrl = image.filename;
+      }
       product.title = title,
         product.price = price,
-        product.imageUrl = imageUrl,
         product.description = description
       product.save()  //if save is passed to an existing product, it won't create a new product; instead it will update it.
         .then(() => {
@@ -159,7 +186,13 @@ exports.getProducts = (req, res, next) => {
 
 exports.deleteProduct = (req, res, next) => {
   const id = req.body.id;
-  Product.deleteOne({ _id: id, userId: req.session.user._id })
+  Product.findById(id).then(prod => {
+    if (!prod) {
+      return next(new Error('No Product found'));
+    }
+    fileHelper.deleteFile(path.join('images', prod.imageUrl));
+    return Product.deleteOne({ _id: id, userId: req.session.user._id })
+  })
     .then(result => {
       res.redirect('../admin/products');
     })
